Removed old exports and tests.
