--- R/lambdaNR.R
@@ -1,31 +1,0 @@
-#' Solves the inner optimization problem of an EL maximization problem.
-#'
-#' @template arg-G
-#' @template arg-lambda0
-#' @param weights Optional length-`n_obs` vector of weights for weighted Newton-Raphson algorithm for right-censored data.
-#' @template args-lambda_precision
-#' @template arg-support
-#' @template arg-verbose
-#' @example examples/lambdaNR.R
-#' @return Vector of length `n_eq` corresponding to the solution of the optimization problem.
-#' @export lambdaNR
-lambdaNR <- function(G, 
-                     lambda0 = rep(0, ncol(G)),
-                     weights = NULL, 
-                     support = FALSE,
-                     max_iter = 100, 
-                     rel_tol = 1e-7, 
-                     verbose = FALSE) {
-  
-  if (is.null(weights)) {
-    lambda <- .LambdaNR(G = t(G), lambda0 = lambda0,
-                        max_iter = max_iter, rel_tol = rel_tol, 
-                        support = support, verbose = verbose)
-  }
-  else {
-    lambda <- .LambdaNRCens(G = t(G), weights = weights, 
-                            max_iter = max_iter, rel_tol = rel_tol, 
-                            support = support, verbose = verbose)
-  }
-  return(lambda)
-}
