--- man/lambdaNR.Rd
@@ -1,54 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/lambdaNR.R
-\name{lambdaNR}
-\alias{lambdaNR}
-\title{Solves the inner optimization problem of an EL maximization problem.}
-\usage{
-lambdaNR(
-  G,
-  lambda0 = rep(0, ncol(G)),
-  weights = NULL,
-  support = FALSE,
-  max_iter = 100,
-  rel_tol = 1e-07,
-  verbose = FALSE
-)
-}
-\arguments{
-\item{G}{\code{nObs x nEqs} constraint matrix in an empirical likelihood framework, where \code{nObs} is the number of observations and \code{nEqs} is the number of constraints.}
-
-\item{lambda0}{A numeric vector of length \code{n_eqs} as the initial value of lambda.}
-
-\item{weights}{Optional length-\code{n_obs} vector of weights for weighted Newton-Raphson algorithm for right-censored data.}
-
-\item{support}{A boolean indicating whether support correction should be conducted.}
-
-\item{max_iter}{Integer scalar for maximum number of Newton-Raphson steps.}
-
-\item{rel_tol}{Positive scalar for relative tolerance of Newton-Raphson steps.}
-
-\item{verbose}{Logical scalar for whether to display number of steps and tolerance criterion when algorithm terminates.}
-}
-\value{
-Vector of length \code{n_eq} corresponding to the solution of the optimization problem.
-}
-\description{
-Solves the inner optimization problem of an EL maximization problem.
-}
-\examples{
-n <- 20
-p <- 8
-G <- matrix(rnorm(n*p),n,p) # randomly generated G
-max_iter <- 100
-rel_tol <- 1e-5
-
-# without support correction
-lambdaNR(G = G,
-         max_iter = max_iter, rel_tol = rel_tol, support = FALSE, 
-         verbose = FALSE)
-
-# with support correction
-lambdaNR(G = G,
-         max_iter = max_iter, rel_tol = rel_tol, support = TRUE, 
-         verbose = FALSE)
-}
