+++ tests/testthat.R
@@ -1,7 +1,7 @@
 library(testthat)
 
 if(!identical(Sys.getenv("NOT_CRAN"), "true")) {
-  set.seed(230)
+  set.seed(123)
 }
 
 test_check("flexEL")
