+++ tests/testthat/test-mr_evalG.R
@@ -5,7 +5,7 @@ source("reg_models.R")
 # library(testthat) # not loaded automatically
 context("mr_evalG")
 
-ntest <- 50
+ntest <- 5
 
 test_that("mr_evalG_R == mr_evalG_cpp", {
   for(ii in 1:ntest) {
